+++ src/app/api/sessions/conflicts/resolve/route.ts
