+++ src/lib/network/safe-api-call.ts
@@ -22,6 +22,7 @@ export async function safeApiCall<T>(
     options?: RequestInit,
     validationSchema?: z.ZodType<T>,
 ): Promise<Result<T, NetworkError>> {
+    console.log('Making API call to:', path);
     let response: Response;
     try {
         response = await fetch(constructUrl(path), {
@@ -38,9 +39,11 @@ export async function safeApiCall<T>(
     }
 
     const responseBody = await readJson(response);
+    console.log('API response body:', responseBody);
 
     if (!response.ok) {
         const parsedError = backendErrorSchema.safeParse(responseBody);
+        console.log('Parsed error:', parsedError);
         return err({
             kind: statusToNetworkErrorKind(response.status),
             status: response.status,
@@ -50,6 +53,7 @@ export async function safeApiCall<T>(
 
     if (validationSchema) {
         const parsedData = validationSchema.safeParse(responseBody);
+        console.log('Parsed data:', parsedData);
         if (!parsedData.success) return err({ kind: 'client' });
         return ok(parsedData.data);
     }
